requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
dataclasses-json==0.6.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.30
langchain-text-splitters==0.3.5
langsmith==0.2.11
Levenshtein==0.26.1
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib-inline==0.1.7
mistune==3.1.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
networkx==3.2.1
numpy==1.26.4
orjson==3.10.15
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-Levenshtein==0.26.1
python-telegram-bot==20.0
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rpds-py==0.22.3
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.37
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.48.0
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yarl==1.18.3
zipp==3.21.0
